Move readers to the storage package
