+++ storage/readers.go
@@ -1,4 +1,4 @@
-package disk
+package storage
 
 import (
 	"io"
